+++ src/types.ts
@@ -35,6 +35,7 @@ import type { AnyWSLocalHook } from './ws/types'
 import type { WebSocketHandler } from './ws/bun'
 
 import type { Instruction as ExactMirrorInstruction } from 'exact-mirror'
+import { BunHTMLBundlelike } from './universal/types'
 
 export type IsNever<T> = [T] extends [never] ? true : false
 
@@ -1063,7 +1064,8 @@ export type MacroToContext<
 											// @ts-expect-error type is checked in key mapping
 											Value['resolve']
 										>
-									> & MacroToContext<
+									> &
+									MacroToContext<
 										MacroFn,
 										// @ts-ignore trust me bro
 										Pick<
@@ -1143,6 +1145,25 @@ type InlineResponse =
 	| AnyElysiaCustomStatusResponse
 	| ElysiaFile
 	| Record<any, unknown>
+	| BunHTMLBundlelike
+
+type LastOf<T> =
+	UnionToIntersect<T extends any ? () => T : never> extends () => infer R
+		? R
+		: never
+
+type Push<T extends any[], V> = [...T, V]
+
+type TuplifyUnion<
+	T,
+	L = LastOf<T>,
+	N = [T] extends [never] ? true : false
+> = true extends N ? [] : Push<TuplifyUnion<Exclude<T, L>>, L>
+
+export type Tuple<
+	T,
+	A extends T[] = []
+> = TuplifyUnion<T>['length'] extends A['length'] ? [...A] : Tuple<T, [T, ...A]>
 
 export type InlineHandler<
 	Route extends RouteSchema = {},
@@ -1177,7 +1198,13 @@ export type InlineHandler<
 								| (Route['response'] extends {
 										200: any
 								  }
-										? Route['response'][200]
+										?
+												| Route['response'][200]
+												| ElysiaCustomStatusResponse<
+														200,
+														Route['response'][200],
+														200
+												  >
 										: unknown)
 								// This could be possible because of set.status
 								| Route['response'][keyof Route['response']]
@@ -1732,7 +1759,7 @@ export type GuardLocalHook<
 	AfterHandle extends MaybeArray<AfterHandler<any, any>>,
 	ErrorHandle extends MaybeArray<ErrorHandler<any, any, any>>,
 	GuardType extends GuardSchemaType = 'standalone',
-	AsType extends LifeCycleType = 'local',
+	AsType extends LifeCycleType = 'local'
 > = (Input extends any ? Input : Prettify<Input>) & {
 	/**
 	 * @default 'override'
@@ -2015,6 +2042,17 @@ export interface EmptyRouteSchema {
 	response: unknown
 }
 
+export interface UnknownRouteSchema<
+	Params = { [name: string]: string | undefined }
+> {
+	body: unknown
+	headers: { [name: string]: string | undefined }
+	query: { [name: string]: string | undefined }
+	params: Params
+	cookie: {}
+	response: unknown
+}
+
 type Extract200<T> = T extends AnyElysiaCustomStatusResponse
 	?
 			| Exclude<T, AnyElysiaCustomStatusResponse>
@@ -2190,10 +2228,10 @@ type PartialIf<T, Condition extends boolean> = Condition extends true
 // Exclude return error()
 export type ExcludeElysiaResponse<T> = PartialIf<
 	Exclude<Awaited<T>, AnyElysiaCustomStatusResponse> extends infer A
-		? IsNever<A> extends true
+		? IsNever<A & {}> extends true
 			? {}
 			: // Intersect all union and fallback never to {}
-				UnionToIntersect<A & {}>
+				A & {}
 		: {},
 	undefined extends Awaited<T> ? true : false
 >